Separating and cleaning.
Using the style of function ptr as state from Rob Pike's presentation
on the template lexer in go. Splitting into separate lexer (scan the
markdown and classify by line) and parser (build blocks by interpreting
the line-level structure).

These will end up in different files (need to experiment with go packaging
and layout) and with a channel between them instead of a big array of
classified lines. This will allow them to run in separate goroutines.