Frontmatter work.
Attributes of the header are passed in the Block for the BigHeading which is now used to
build the page header. This pulls in the target style file as an extra on top of the
base style file.

The example now renders correctly other than:
* The bibitem is missing the book item.
* The image and video paths are not real.
* Need the regex search/replace to overwrite inline style.
* Font size for quotes seems overly large.
* Looks like the wrong label for deflist items.

Then a separate slides.css is needed...