Longform handling.
Placed a buffer in the html render that acts as multiple output channels
while building the slide. This gives a primary / secondary / longform
channel and allows the state changes to shift between them. The old emit
of html is directed through this buffer and the entire slide is generated
and inserted into the output stream as a finalise step.