Cleaned up similar handler cases.
There was some cut and paste between different filetype extensions in
order to handle the mimetype differences. Put the mapping into a
constant map instead and collapsed the cases together.

Added a new case to handle .svg files for Linux 08.

Witnessed a disturbing crash while running the app locally in a specially
crafted fake environment. It looks like a coroutine leak that has caused
the runtime to run out of memory. I have no repeatable test case, it was
acting a bit weird for an hour before it crashed. Bit worrying.