Blog working well.
The application is up and running, uses a custom renderer in the blog
main package rather than the one in the rst package. This has necessitated
converting the Block interface to something exported - in golang this
changes all of the identifiers.

I've been through any changed most of the points that I could find in
the rst package, but the other apps still need testing before this is
pushed up to the server.