Fixing redirects
If we hit a directory name then add index.html and send back a 302
redirect. This means that sub-directory urls for images will then
work correctly.

If we use the internal URL given to the backend then we bounce to a
URL that does not exist on the nginx host, so I've added nginx
config that copies the original URI as X-Forwarded-URL which is
used here to build the redirect target.