Haha, trimmed the rest of it away.
The facebook auth-flow is subtly different from the flow in the oidc code so
I've ended up reimplementing the common core.

This works against google. The login cookie is a HMACed session key. We need
to keep some state as the access token is confidential and cannot be shared
via the cookie without encrypting it first.