Comments.
Use a web-socket from the client to the server. Connected when the user starts typing
in the comment textare. This is only visible when logged in and the login cookie is
used to authenticate the incoming ws connection is valid.

Simple protocol on the ws with an action and body. The text area contents is sent as
a preview message to get a html-rendered version of the input, this allows live updates
of the preview.

On post a post message is sent with the text area so that it can be stored. Need to
implement similar json wrapping for protocol in the other direction. On storage of
a new post fire a message to the client to redraw the comments div.