Comments stored in Posts.
Posts are now read through the cache properly. Before, only the metadata
was cached to speed up the index. Now we cache the rendered html as well.

Comments are not persistent yet - need to serialise the []Comment to a
backing file for each post and reload.

Locking will be necessary before it goes live.