Comments are now persistent.
For now we just blow out a json encoded dump of the Post.Comments array. This is
plenty fast given the size of the data. Each dump is given a monotonically
increasing index and when rendering the page we just pick the highest.

Only thing left before deployment is locking around the file-system access to
preserve consistency. This will be more important if I cleanup - on a cache
Lookup that sucessfully reads the highest index we could delete the lower ones
as old data...