Scrubbing passwords / Switching repos
This code was originally in a mega-repo that was not meant for public release.
I'm pulling this code out into a separate repo to put on github. The history
of the rst parser is that it was not originally written for this blog code.
It is a reimplementation of the python parser that I wrote for lecture slides,
that I've ended up repurposing for this.

In parallel to this I will rewrite the new repo history to scrub passwords so
that I can release it. Obviously it never really worked that way. This change
is a fix for the censored passwords.

Check the local credentials properly - there is a hash of the password with
unique salt in the source. If you use this code yourself then please switch
this code to be your local user account - the print will dump out the salted
hash of whatever you type in to make it easier to setup. It is good practice
to change the salt as well.

The OAauth2 shared secrets and clientIDs have been scrubbed in a similar way.
All old versions will be broken (as the providers array just contains the
"CENSORED" string) and this path will initialise the secrets from a secrets.json
file in the same directory as the executable, this file should have the
following syntax:

{ GoogleID: "google application id",
  GoogleSecret: "the shared secret" }

You can find these values in your Google dashboard after you create an app
for the blog.